+++ src/controllers/categories.js
@@ -3,17 +3,26 @@ const _ = require('lodash');
 const uuid = require('uuid');
 const bunyan = require('bunyan');
 const log = bunyan.createLogger({ name: 'book-library-service' });
-const {User} = require('../postgres/models');
+const { User, Book } = require('../postgres/models');
 const Auth = require('./auth');
+const BookApi = require('./book');
 
 module.exports = function (Category) {
   const Authenticate = Auth(User);
+  const Books = BookApi(Book);
 
   const findCategories = async (req, res) => {
     const token = req.headers['x-access-token'] || req.headers['authorization'];
+    // console.log(req.query);
+    const [page, pageSize ] = [ 1, 5]; //[req.query.page, req.query.pageSize] ||;    
+    const offset = page * pageSize;
+    const limit = offset + pageSize;
     try {
       await Authenticate.verifyToken(token);
-      const category = await Category.findAll().then((results) => results).catch(() => []);
+      const category = await Category.findAll({limit, offset})
+        .then((results) => results)
+        .catch(() => []);
+
       res.status(200).send({
         message: 'found category',
         category
@@ -29,6 +38,7 @@ module.exports = function (Category) {
     try {
       await Authenticate.verifyToken(token);
       let { name, description } = req.body;
+      name = _.capitalize(name); description = _.capitalize(description);
       let data = { name, categoryId: uuid.v4(), description };
       let [category, created] = await Category.findOrCreate({
         where: { name }, defaults: data
@@ -97,16 +107,21 @@ module.exports = function (Category) {
     try {
       await Authenticate.verifyToken(token);
       const { id } = req.params;
-      Category.findOne({
+      await Category.findOne({
         where: { categoryId: id }
-      }).then((category) => {
-        category.destroy({ force: true });
-      }).catch((err) => {
-        throw new Error(err);
+      }).then(async (category) => {
+        const { name } = category || '';
+        let book = await Books.getBookByCategory(name);
+        if (name && _.isEmpty(book)) { 
+          category.destroy({ force: true }); 
+          res.status(200).send({ message: 'Category deleted!' });
+        } else {
+          throw new Error(`Failed to delete category with id:${id}`);
+        }
       });
     } catch (error) {
-      log.warn(error);
-      res.status(500).send('Error occurred');
+      // log.warn(error);
+      res.status(500).send(error.message);
     }
   };
